Move split tunnel request handling to own module
